Add files via upload
Bibliography file - It is linked to the Zotero Multispecies Model repository. Just make sure that this file is in your working directory. You might need to download Better Bib(La)TeX as a Zotero add on (you download it in Zotero). That way you can set Zotero to automatically update this file as you add papers to the repository and will auto download.